+++ examples/complete/data.tf
@@ -1,8 +1,19 @@
-data "aws_caller_identity" "current" {}
 
-data "aws_region" "current" {}
+data "aws_subnets" "private" {
+  filter {
+    name   = "tag:Name"
+    values = ["${local.supporting_resources_name}*.pri.*"]
+  }
+}
 
+data "aws_vpc" "supporting" {
+  filter {
+    name   = "tag:Name"
+    values = [local.supporting_resources_name]
+  }
+}
 
-data "aws_availability_zones" "available" {
-  state = "available"
+data "aws_subnet" "private" {
+  for_each = toset(data.aws_subnets.private.ids)
+  id       = each.value
 }