+++ examples/complete/outputs.tf
@@ -1,6 +1,1 @@
-output "complete_cluster" {
-  description = "Example output for complete cluster."
-  value = [
-    module.complete_cluster,
-  ]
-}
+# Empty